+++ colorFromPairNumber.js
@@ -1,6 +1,4 @@
-var MajorColors = require("./minor_major_color").MajorColors;
-var MinorColors = require("./minor_major_color").MinorColors;
-
+const {MajorColors,MinorColors} = require("./minor_major_color");
 
 function GetColorFromPairNumber(pairNumber) {
     let colorPair = {};
