--- testNumbertoPair.js
@@ -1,11 +1,0 @@
-const expect = require('chai').expect;
-var GetColorFromPairNumber = require("./colorFromPairNumber.js");
-
-function testNumberToPair(number, expectedMajor, expectedMinor) {
-    const pairOfColors = GetColorFromPairNumber(number);
-    console.log(`${number} = ${expectedMajor} ${expectedMinor}`);
-    expect(pairOfColors.major).equals(expectedMajor);
-    expect(pairOfColors.minor).equals(expectedMinor);
-}
-
-module.exports = testNumberToPair;
