+++ testing.js
@@ -1,0 +1,24 @@
+const expect = require('chai').expect;
+const GetPairNumberFromColors = require("./pairNumberFromColors.js");
+const GetColorFromPairNumber = require("./colorFromPairNumber.js");
+
+function testColorToNumber(majorColor, minorColor, expectedNumber) {
+    const pairNumber = GetPairNumberFromColors(majorColor, minorColor);
+    console.log(`${majorColor} ${minorColor} = ${pairNumber}`);
+    expect(pairNumber).to.equal(expectedNumber);
+}
+
+
+
+function testNumberToPair(number, expectedMajor, expectedMinor) {
+    const pairOfColors = GetColorFromPairNumber(number);
+    console.log(`${number} = ${expectedMajor} ${expectedMinor}`);
+    expect(pairOfColors.major).equals(expectedMajor);
+    expect(pairOfColors.minor).equals(expectedMinor);
+}
+
+console.log("color coder");
+testNumberToPair(4, "white", "brown");
+testNumberToPair(25, "violet", "slate");
+testColorToNumber("black", "brown", 14);
+testColorToNumber("yellow", "green", 18);
